web/pyproject.toml
[tool.poetry]
name = "web"
version = "0.1.0"
description = ""
authors = ["mbashem <mohammedbasem.patwary@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.5"
django-allauth = "^0.51.0"
python-dotenv = "^0.20.0"
djongo = "^1.3.6"
djangorestframework = "^3.13.1"
django-cors-headers = "^3.13.0"
pip = "^22.1.2"
install = "^1.3.5"
pymongo = "3.12.3"
dnspython = "^2.2.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[tool.poetry.scripts]
server = "web.scripts:django_run_server"
django = "web.scripts:django_run_manage"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
